Prepare for separation of js from php; header.inc.php done
