+++ .gitignore
@@ -53,3 +53,4 @@ web.config
 *~
 .php-cs-fixer.cache
 .phpunit.result.cache
+public/dist/
