+++ composer.json
@@ -56,6 +56,7 @@
         "doctrine/cache": "^2",
         "gettext/gettext": "^5",
         "gettext/translator": "^1.2",
+        "idleberg/vite-manifest": "^1.0",
         "james-heinrich/getid3": "dev-master",
         "jwilsson/spotify-web-api-php": "^6",
         "krixon/xbmc-php-rpc": "dev-master",
