+++ docs/examples/update_from_git.sh
@@ -34,5 +34,6 @@ php bin/cli run:updateCatalog -t
 ### You don't always need to do this but some people might want to keep composer packages updated here
 #composer install --prefer-source --no-interaction
 
-### NPM install is now required to copy all the javascript packages
+### NPM is now required to handle all the javascript packages
 npm install
+npm run build
