+++ package.json
@@ -2,6 +2,7 @@
   "name": "ampache",
   "version": "1.0.0",
   "description": "Packages used by Ampache",
+  "type": "module",
   "private": "true",
   "dependencies": {
     "@material-symbols/svg-400": "^0.16.0",
@@ -19,6 +20,8 @@
     "wunderbaum": "0.8.2"
   },
   "scripts": {
+    "dev": "vite",
+    "build": "vite build",
     "postinstall": "npm-run-all copy:*",
     "copy:bootstrap": "copyfiles --up 3 node_modules/bootstrap/dist/**/* public/lib/components/bootstrap",
     "copy:contextmenu": "copyfiles --up 3 node_modules/jquery-contextmenu/dist/* public/lib/components/jquery-contextmenu",
@@ -50,6 +53,7 @@
   "homepage": "https://github.com/ampache/ampache#readme",
   "devDependencies": {
     "copyfiles": "^2.4.1",
-    "npm-run-all": "^4.1.5"
+    "npm-run-all": "^4.1.5",
+    "vite": "^5.2.10"
   }
 }
