+++ public/templates/show_html5_player_headers.inc.php
@@ -41,9 +41,6 @@ if ($iframed || $isShare) { ?>
     <script src="<?php echo $web_path; ?>/lib/components/jquery/jquery.min.js"></script>
     <script src="<?php echo $web_path; ?>/lib/components/jquery-ui/jquery-ui.min.js"></script>
     <script src="<?php echo $web_path; ?>/lib/components/js-cookie/js.cookie.js"></script>
-    <script src="<?php echo $web_path; ?>/lib/javascript/base.js"></script>
-    <script src="<?php echo $web_path; ?>/lib/javascript/ajax.js"></script>
-    <script src="<?php echo $web_path; ?>/lib/javascript/tools.js"></script>
     <script>
         var jsAjaxServer = "<?php echo $ajaxUriRetriever->getAjaxServerUri(); ?>";
         var jsAjaxUrl = "<?php echo $ajaxUriRetriever->getAjaxUri(); ?>";
