+++ public/templates/show_registration_confirmation.inc.php
@@ -47,8 +47,6 @@ $_SESSION['login'] = true; ?>
                 <a href="<?php echo $web_path; ?>"><h1 id="headerlogo"></h1></a>
             </div>
             <script src="<?php echo $web_path; ?>/lib/components/jquery/jquery.min.js"></script>
-            <script src="<?php echo $web_path; ?>/lib/javascript/base.js"></script>
-            <script src="<?php echo $web_path; ?>/lib/javascript/ajax.js"></script>
             <div id="content">
                 <div id="guts">
 <?php $url = $web_path . '/login.php';
