+++ public/templates/show_user_registration.inc.php
@@ -53,8 +53,6 @@ $_SESSION['login'] = true; ?>
 
 <body id="registerPage">
     <script src="<?php echo $web_path; ?>/lib/components/jquery/jquery.min.js"></script>
-    <script src="<?php echo $web_path; ?>/lib/javascript/base.js"></script>
-    <script src="<?php echo $web_path; ?>/lib/javascript/ajax.js"></script>
 
     <div id="maincontainer">
         <div id="header">
